pyproject.toml
[project]
name = "multi-agent-research"
version = "0.1.0"
description = "Multi-agent system using LangChain for collaborative research, fact-checking, and content creation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "python-dotenv>=1.0.0",
]
